inference: re-architect entry points for more clarity
The main functional changes are:
 - More nearly separates the optimization steps from the inference loop
   (still some interleaving, due to wanting to cache work results eagerly)
 - Doesn't run `widen_all_const` on the IR before returning to the user
   during reflection (e.g. code-typed). This allows the user to run
   additional passes on it outside of Julia codegen,
   or simply to better see what inference knows.
 - There's now a flag in `Compiler.Params` for whether the global
   inference cache is applicable and should be used (read / write).
   Although this flag is currently not respected, however.
 - Remove some dead code vestiges still left over from the old optimizer.
