+++ base/compiler/typeutils.jl
@@ -10,8 +10,7 @@ function rewrap(@nospecialize(t), @nospecialize(u))
     return rewrap_unionall(t, u)
 end
 
-const _TYPE_NAME = Type.body.name
-isType(@nospecialize t) = isa(t, DataType) && (t::DataType).name === _TYPE_NAME
+isType(@nospecialize t) = isa(t, DataType) && t.name === _TYPE_NAME
 
 # true if Type{T} is inlineable as constant T
 # requires that T is a singleton, s.t. T == S implies T === S
@@ -27,19 +26,9 @@ function issingletontype(@nospecialize t)
     iskindtype(typeof(t)) || return true # non-types are always compared by egal in the type system
     isconcretetype(t) && return true # these are also interned and pointer comparable
     if isa(t, DataType) && t.name !== Tuple.name && !isvarargtype(t) # invariant DataTypes
-        return all(p -> issingletontype(p), t.parameters)
+        return all(@nospecialize(p) -> issingletontype(p), t.parameters)
     end
     return false
-end
-
-iskindtype(@nospecialize t) = (t === DataType || t === UnionAll || t === Union || t === typeof(Bottom))
-isconcretedispatch(@nospecialize t) = isconcretetype(t) && !iskindtype(t)
-
-# equivalent to isdispatchtuple(Tuple{v}) || v == Union{}
-# and is thus perhaps most similar to the old (pre-1.0) `isleaftype` query
-function isdispatchelem(@nospecialize v)
-    return (v === Bottom) || (v === typeof(Bottom)) ||
-        isconcretedispatch(v) || (isType(v) && !has_free_typevars(v))
 end
 
 argtypes_to_type(argtypes::Array{Any,1}) = Tuple{anymap(widenconst, argtypes)...}
@@ -59,8 +48,6 @@ function valid_tparam(@nospecialize(x))
     end
     return isa(x, Symbol) || isbitstype(typeof(x))
 end
-
-has_free_typevars(@nospecialize(t)) = ccall(:jl_has_free_typevars, Cint, (Any,), t) != 0
 
 # return an upper-bound on type `a` with type `b` removed
 # such that `return <: a` && `Union{return, b} == Union{a, b}`
@@ -96,7 +83,19 @@ _typename(union::UnionAll) = _typename(union.body)
 _typename(a::DataType) = Const(a.name)
 
 function tuple_tail_elem(@nospecialize(init), ct)
-    return Vararg{widenconst(foldl((a, b) -> tmerge(a, tvar_extent(unwrapva(b))), init, ct))}
+    # FIXME: this is broken: it violates subtyping relations and creates invalid types with free typevars
+    tmerge_maybe_vararg(@nospecialize(a), @nospecialize(b)) = tmerge(a, tvar_extent(unwrapva(b)))
+    return Vararg{widenconst(foldl(tmerge_maybe_vararg, init, ct))}
+end
+
+function countunionsplit(atypes)
+    nu = 1
+    for ti in atypes
+        if isa(ti, Union)
+            nu *= unionlen(ti::Union)
+        end
+    end
+    return nu
 end
 
 # take a Tuple where one or more parameters are Unions
