+++ stdlib/InteractiveUtils/src/codeview.jl
@@ -2,19 +2,13 @@
 
 # displaying type warnings
 
-function code_warntype_legacy_ir(io::IO, src::Core.CodeInfo, rettype)
-    function slots_used(ci, slotnames)
-        used = falses(length(slotnames))
-        scan_exprs!(used, ci.code)
-        return used
-    end
-
-    function scan_exprs!(used, exprs)
+function code_warntype_legacy_ir(io::IO, src::Core.CodeInfo, @nospecialize(rettype))
+    function scan_expr_uses!(used, exprs::Vector{Any})
         for ex in exprs
             if isa(ex, Core.Slot)
                 used[ex.id] = true
             elseif isa(ex, Expr)
-                scan_exprs!(used, ex.args)
+                scan_expr_uses!(used, ex.args)
             end
         end
     end
@@ -22,7 +16,9 @@ function code_warntype_legacy_ir(io::IO, src::Core.CodeInfo, rettype)
     emph_io = IOContext(io, :TYPEEMPHASIZE => true)
     println(emph_io, "Variables:")
     slotnames = Base.sourceinfo_slotnames(src)
-    used_slotids = slots_used(src, slotnames)
+    # compute the set of ssa values that are used
+    used_slotids = falses(length(slotnames))
+    scan_expr_uses!(used_slotids, src.code)
     for i = 1:length(slotnames)
         if used_slotids[i]
             print(emph_io, "  ", slotnames[i])
@@ -44,8 +40,8 @@ function code_warntype_legacy_ir(io::IO, src::Core.CodeInfo, rettype)
     print(emph_io, '\n')
 end
 
-function warntype_type_printer(io, ty)
-    if ty !== Union{} && (!isdispatchtuple(Tuple{ty}) || ty == Core.Box)
+function warntype_type_printer(io::IO, @nospecialize(ty))
+    if ty isa Type && (!Base.isdispatchelem(ty) || ty == Core.Box)
         if ty isa Union && Base.is_expected_union(ty)
             Base.emphasize(io, "::$ty", Base.warn_color()) # more mild user notification
         else
@@ -54,6 +50,7 @@ function warntype_type_printer(io, ty)
     else
         Base.printstyled(io, "::$ty", color=:cyan)
     end
+    nothing
 end
 
 """
@@ -71,21 +68,25 @@ in verbose mode, showing all available information (rather than applying
 the usual heuristics).
 See [`@code_warntype`](@ref man-code-warntype) for more information.
 """
-function code_warntype(io::IO, f, @nospecialize(t); verbose_linetable=false)
+function code_warntype(io::IO, @nospecialize(f), @nospecialize(t); verbose_linetable=false)
     for (src, rettype) in code_typed(f, t)
         if src.codelocs === nothing
             code_warntype_legacy_ir(io, src, rettype)
         else
-            print(io, "Body"); warntype_type_printer(io, rettype); println(io);
+            print(io, "Body")
+            warntype_type_printer(io, rettype)
+            println(io)
             # TODO: static parameter values
             ir = Core.Compiler.inflate_ir(src, Core.svec())
             Base.IRShow.show_ir(io, ir, warntype_type_printer;
-                                argnames=Base.sourceinfo_slotnames(src), verbose_linetable=verbose_linetable)
+                                argnames = Base.sourceinfo_slotnames(src),
+                                verbose_linetable = verbose_linetable)
         end
     end
     nothing
 end
-code_warntype(f, @nospecialize(t); kwargs...) = code_warntype(stdout, f, t; kwargs...)
+code_warntype(@nospecialize(f), @nospecialize(t); kwargs...) =
+    code_warntype(stdout, f, t; kwargs...)
 
 import Base.CodegenParams
