+++ stdlib/REPL/src/REPLCompletions.jl
@@ -321,7 +321,7 @@ function get_type_call(expr::Expr)
     m = first(mt)
     # Typeinference
     params = Core.Compiler.Params(world)
-    return_type = Core.Compiler.typeinf_type(m[3], m[1], m[2], true, params)
+    return_type = Core.Compiler.typeinf_type(m[3], m[1], m[2], params)
     return_type === nothing && return (Any, false)
     return (return_type, true)
 end
